Add Sprite classes, update Animation classes for offsets and height and width.

+++ src/main/java/org/abberkeep/gameframework/motion/EightWayMotion.java
@@ -120,12 +120,12 @@ public class EightWayMotion implements Motion {
    }
 
    @Override
-   public float getHeight() {
+   public int getHeight() {
       return animations[index].getHeight();
    }
 
    @Override
-   public float getWidth() {
+   public int getWidth() {
       return animations[index].getWidth();
    }
 
@@ -151,7 +151,7 @@ public class EightWayMotion implements Motion {
    }
 
    @Override
-   public void setSize(float width, float height) {
+   public void setSize(int width, int height) {
       for (Animation animation1 : animations) {
          animation1.setSize(width, height);
       }
